LayoutTests/platform/chromium-mac/platform/chromium/fast/forms/calendar-picker/calendar-picker-f4-key-expected.txt

Tests if calendar picker key bindings work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
FAIL document.getElementById("mock-page-popup") should be non-null. Was null
PASS successfullyParsed is true

TEST COMPLETE
